// apps/api/src/shared/zod/schemas/games.schema.ts
export const RetroachievementsSchema = z.object({
  gameId: z.number(),
  consoleId: z.number(),
});

export const IGDBSchema = z.object({
  gameId: z.number(),
  total_rating: z.number().nullable().optional(),
  total_rating_count: z.number().nullable().optional(),
  aggregated_rating: z.number().nullable().optional(),
  aggregated_rating_count: z.number().nullable().optional(),
  rating: z.number().nullable().optional(),
  rating_count: z.number().nullable().optional(),
  hypes: z.number().nullable().optional(),
  screenshotsCount: z.number().nullable().optional(),
  artworksCount: z.number().nullable().optional(),
  status: z.number().nullable().optional(),
  version_parent: z.number().nullable().optional(),
  parent_game: z.number().nullable().optional(),
  url: z.string().nullable().optional(),
  genres: z.number().array().optional(),
  keywords: z.number().array().optional(),
  themes: z.number().array().optional(),
  modes: z.number().array().optional(),
  websites: z.number().array().optional(),
  release_dates: z.number().array().optional(),
  platforms: z.number().array().optional(),
  involved_companies: z.number().array().optional(),
  game_type: z.number().nullable().optional(),
  cover: z.number().array().optional(),
  screenshots: z.number().array().optional(),
  artworks: z.number().array().optional(),
  franchises: z.number().array().optional(),
  videos: z.number().array().optional(),
  alternative_names: z.number().array().optional(),
  game_engines: z.number().array().optional(),
  player_perspectives: z.number().array().optional(),
  dlcs: z.number().array().optional(),
  expansions: z.number().array().optional(),
  standalone_expansions: z.number().array().optional(),
  bundles: z.number().array().optional(),
  expanded_games: z.number().array().optional(),
  forks: z.number().array().optional(),
  ports: z.number().array().optional(),
  remakes: z.number().array().optional(),
  remasters: z.number().array().optional(),
  similar_games: z.number().array().optional(),
});

export const MultiplayerModeSchema = z.object({
  platformId: z.string().nullable().optional(),
  campaignCoop: z.boolean().optional(),
  dropIn: z.boolean().optional(),
  lanCoop: z.boolean().optional(),
  offlineCoop: z.boolean().optional(),
  offlineCoopMax: z.number().nullable().optional(),
  offlineMax: z.number().nullable().optional(),
  onlineCoop: z.boolean().optional(),
  onlineCoopMax: z.number().nullable().optional(),
  onlineMax: z.number().nullable().optional(),
  splitscreen: z.boolean().optional(),
  splitscreenOnline: z.boolean().optional(),
});

export const AgeRatingSchema = z.object({
  organization: z.string(),
  rating: z.string(),
  synopsis: z.string().optional(),
});

export const ExternalPageSchema = z.object({
  name: z.string().optional(),
  uid: z.string(),
  url: z.string().nullable().optional(),
});

export const RelatedGamesSchema = z.object({
  parent_game: z.string().nullable().optional(),
  dlcs: z.string().array().optional(),
  expansions: z.string().array().optional(),
  standalone_expansions: z.string().array().optional(),
  bundles: z.string().array().optional(),
  expanded_games: z.string().array().optional(),
  forks: z.string().array().optional(),
  ports: z.string().array().optional(),
  remakes: z.string().array().optional(),
  remasters: z.string().array().optional(),
  similar_games: z.string().array().optional(),
});

export const HltbSchema = z.object({
  hltbId: z.string(),
  mainStory: z.number().nullable().optional(),
  mainExtra: z.number().nullable().optional(),
  completionist: z.number().nullable().optional(),
  allStyles: z.number().nullable().optional(),
  coop: z.number().nullable().optional(),
  multiplayer: z.number().nullable().optional(),
  mainStoryCount: z.number().nullable().optional(),
  mainExtraCount: z.number().nullable().optional(),
  completionistCount: z.number().nullable().optional(),
  allStylesCount: z.number().nullable().optional(),
  coopCount: z.number().nullable().optional(),
  multiplayerCount: z.number().nullable().optional(),
  reviewScore: z.number().nullable().optional(),
  imageUrl: z.string().nullable().optional(),
  platforms: z.string().array().optional(),
  releaseYear: z.number().nullable().optional(),
  similarity: z.number().nullable().optional(),
  alias: z.string().nullable().optional(),
  type: z.string().nullable().optional(),
  sourceName: z.string().optional(),
  updatedAt: z.string(),
});

export const CompanySchema = z.object({
  name: z.string(),
  developer: z.boolean().default(false),
  porting: z.boolean().default(false),
  publisher: z.boolean().default(true),
  supporting: z.boolean().default(false),
});
// ...
export const ReleaseDateSchema = z.object({
  date: z.number(),
  human: z.string(),
  month: z.number(),
  year: z.number(),
  platformId: z.string(),
  region: z.coerce.number(),
});

export const GameSchema = z.object({
  _id: z.string(),
  slug: z.string(),
  name: z.string(),
  type: z.string(),
  cover: z.string().nullable(),
  storyline: z.string().optional(),
  summary: z.string().optional(),
  modes: z.string().array().optional(),
  genres: z.string().array().optional(),
  keywords: z.string().array().optional(),
  themes: z.string().array().optional(),
  screenshots: z.string().array().optional(),
  artworks: z.string().array().optional(),
  franchises: z.string().array().optional(),
  videos: z.string().array().optional(),
  alternative_names: z.string().array().optional(),
  companies: CompanySchema.array().optional(),
  websites: z.string().array().optional(),
  first_release: z.number().nullable().optional(),
  release_dates: ReleaseDateSchema.array().optional(),
  platformIds: z.string().array(),
  status: z.string().nullable().optional(),
  versionTitle: z.string().nullable().optional(),
  game_engines: z.string().array().optional(),
  player_perspectives: z.string().array().optional(),
  multiplayer_modes: MultiplayerModeSchema.array().optional(),
  ageRatings: AgeRatingSchema.array().optional(),
  languages: z.string().array().optional(),
  externalPages: ExternalPageSchema.array().optional(),
  relatedGames: RelatedGamesSchema.optional(),
  characters: CharacterSchema.array().optional(),
  averageRating: z.number().nullable().optional(),
  ratingsCount: z.number().nullable().optional(),
  isStopParsingPictures: z.boolean().optional(),
  isStopParsing: z.boolean().optional(),
  isCustom: z.boolean().optional(),
  retroachievements: RetroachievementsSchema.array().optional(),
  igdb: IGDBSchema.optional(),
  hltb: HltbSchema.optional(),
  createdAt: z.string(),
  updatedAt: z.string(),
});
// ...
export const AddGameRequestSchema = GameSchema.omit({
  _id: true,
  updatedAt: true,
  createdAt: true,
  averageRating: true,
  ratingsCount: true,
  isCustom: true,
});